[TRACE] Fix usage of '__VA_ARGS__' in variadic macros
The recommended way to use '__VA_ARGS__' in variadic macros, is to
prepend it with '##', to avoid problems with some compilers when the
macro is invoked with zero variable arguments (see [1]).

Not prepending '__VA_ARGS__' with '##' may result on build errors with
some compilers. This is actually happening with calls like:

    TRACE_L1("Some message with no variable arguments");

This way, replace all instances of '__VA_ARGS__' with '##__VA_ARGS__'.
Successfully build and runtime tested.

[1] https://gcc.gnu.org/onlinedocs/cpp/Variadic-Macros.html

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>