[OCDM] Fix missing lock Unlock() in OpenCDMAccessor::Release()
When reaching a ref count of zero, OpenCDMAccessor::Release() has a
return path that doesn't release the _systemLock, which causes, e.g., a
new instance creation via OpenCDMAccessor::Instance() to deadlock as
soon as it tries to grab the _systemLock again.

Fix this unbalanced use of _systemLock in OpenCDMAccessor::Release() by
making sure all code paths release the lock.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>