[Core RPC] Fix use of_adminLock for the _stubs map
RPC Invoke() calls may cause that another Invoke() is triggered for the
same Administrator instance, as a result and necessary condition for
processing the first Invoke() call.

The second Invoke() call can be served by another minion (thread) if a
plugin as been configured to have more than one, but then a deadlock
will still happen, because the second Invoke() will be blocked when
grabbing the _adminLock (which is taken by the first Invoke()) and the
first Invoke (holding the lock) is waiting on the second Invoke() to
complete.

The _adminLock is being used to protect the _stubs and the _proxy maps.
For the _stubs case and the Invoke() call, it's only required to hold it
for finding the stub, not while calling the actual handler during the
Handle() call. This way, to avoid the deadlock in Invoke(), just hold
the lock for finding the stub and release it before the Handle() call.
This assumes that the _stubs map is only used for inserting and finding
entries, which is currently true.

While at it, also fix Announce() by protecting it with the _adminLock,
which is required because it is the method that inserts entries in the
_stubs and _proxy map.

Signed-off-by: Ricardo Silva <ricardo.silva@gbtembedded.com>