Make the build environment backwards compatible
There might already be recipies out there that do not set the PLUGIN_COMPOSITOR_SUB_IMPLEMENTATION build variable. This is because there used to be no diffrentiation on Wayland implementations. The default of this value should be set to Westeros as that was the intention. Builds requiring Weston are now possible as well but these builds need to set the PLUGIN_COMPOSITOR_SUB_IMPLEMENTATION to Weston.