RemoteControl: Add new interface to send events
This commit adds a new interface IRemoteControl, This interface
is to be used for registering events from RemoteControl plugin.

To get events to registered clients a Notification class is also
added to IRemoteControl.

A new method is added to IKeyHandler to get the Producer events
from IKeyProducer.

Currently the producer events considered are:

ePairingStarted
ePairingSuccess
ePairingFailed
ePairingTimedout

Signed-off-by: harikrishnan.p <harikris2012@gmail.com>