[PROXY] Depending on the Proxy Object, make sure the proper lifetime management interface is added.
If a COM interface is part of the Proxy object, Lifetime is managed by the IUnknown (IReferenceCounted base) interface, so the ProxyObject should not be created with an additional IReferenceCounted interface. So do not add it to the ProxyObject.
If it a s none IReferenceCounted inheriting object, there is aneed to add this interface to the object. So using type traits make sure teh ProxyObject is created diffrently.