[Process Containers] Made PC work with latest Thunder
There were changes in the way how plugins are started, that made
containers not compile.

Also, now more environment variables are required to be present to
correctly start a WPEProcess (previously only PATH was set, but not
really needed). That uncovered a bug in LXC, where environmental
variables are not inherited from host even if they are requested in a
container config file. That was fixed by implementing this functionality
on Thunder side.

In future it might be fixed from LXC side. Then the fix in Thunder will
be unneccessary, but shouldn't affect the correct LXC solution in a
conflicting way.