[CRASH] Plugins with an IStateControl crash on deactivation.
Plugins with an IStateControl and an open WebController page, might end up in a crash if on one thread, the plugin is de-activiating and on another thread the status for this plugin is requested.
Since it was not loacked, the retrieval of the sate, the plugin might have been considered ACTIVATED, about to get the IStateControl. System is preempted, The plugin is "deactivated", and thus IStateControl
interface became deleted. After continueation of the preemption, the IStateControl, just cleared, get quesries..... Poof !