[IDISPLAY] Make preperations for all the AppEngines to run on a RPI4.
The RaspberryPi4 is aquiring access to the HDMI/Display and the construction of
the surface to draw on differently. This is the abstraction layer to get a surface
created by the hardware and thus needs to be adapted.

Using this adaptions, all AppEngines can aquire an EGL surface to draw on from:

1) WesterOS
2) Weston
3) Nexus
4) RPI0123 DispmanX
5) RPI4 VC6