[RDK-B] Preparations to build for RDK-B.
RDK-B is still relying on an older compiler (4.8) This means some
functionality is missing from the standard libraries. As there
are trials to test Thunder core on the RDK-B, the Thunder core
should avoid using these new standard libraries features.
This commit allows Thunder to be build, using a GCC 4.8 compiler
and associated standard libraries.