[RPI] The RPI client is the only client with a dual function on the Client implementation. Break it.
The lifetime of the C++ object (Compositor::IDisplay::ISurface) should be independent of the COMRPC
object (Exchange::IComposition::IClient) Hence why it should be decoupled. This commit decouples the
lifetinme of these objects and avoids the pure virtual method call if the C++ object is destructed
and the COMPRC object is not yt destructed.