[KEYFLOW] Virtual Input xtend with functionality to selectively dispatch.
To align the functionality offered by Westeros to only forward keys to a surface that has the focus,
the Thunder Key dispatching mechanism needs to expose this functionality in a generic why. The
functionality that was already present for a derived version of the Keyhandler (IPC) has now
been made in a generic way. The VirtualInput object exposes a interator that lists all the
consuming channels. Thise channels can now be enabled/disabled on request (from C++ world)

Next step is to extend this functionality to the Web World. See ThunderNanoServices/InputSwitch
for extensions.