[KEYFLOW] Make sure we can configure whetiher or not the keys will be send to the outputi, by default.
In composition mode, it is required that the App (callsign) receiving the keys, must be determined by the
compositor/mainApp, we would like to start a default new channel to be disabled to receive input events
and only enable it on command. To be able to do so, we need a default settings value. This can now be set
using the following build flag -DKEY_OUTPUT_DISABLED. If this is set, by default new output channels for
receiving keys from the virtualInput will not be enbled and thus control is moved to a higher (Compositor)
layer, to enable this, if required..