[JSONDESERIALIZE] Quotes inside the opaque string needs to be considered quoted as well.
If, during deserialization, the string started to be an opaque string (opening with a '[' or '{') within these structures, the strings inside quotes should be treated as strings inside quotes and not to be parsed on behalf of the opaque structures. So now quotes within opaque structures are treated as quoted as well.