[IEXTERNAL] Remove type from Identifier, make it a settable moduleId.
When more IExternals will be published the type of the IExternal (what the root source of the
device is) is not so important. The device type is abstracted away. From an IExternal level you
can read and/or write a value on it (int32_t).

By removing the the source type of the IExternal, the top 4 bits become available again, together
with the bits 24-27 we free up a whole byte that can be used the "assign" a module identifier to
the IExternalIdentifier. This way, if a plugin collects all IExternals, it is free to assign a
module identifier (based on the interface offering the IExternal) to the IExternal::Identifier.
This will make the Identifier reported by the IExternal unique over multiple ICatalogs.