[DEPENDENCY CLEANUP] Correct the dependencies for interfaces and modules.
Over time some modules/interfaces have inherited incorrect dependencies. For supporting
the default of visibility hidden, changes where made to all Module.h header files. In
this commit we cleanup the usage of the EXTERNAL keyword as well. On linux, the default
for the EXTERNAL macro is "visible". Only on windows we need to toggle this keyword in
case we are building the specific module. Linux requires to change at all, once defined
in Portability.h it is good for all code.

Also some "side-effect" dependencies where dragged in through Interface Header files
like the interface header files in IPlugin. This commit also brings back the dependencies
of these interface header files to com, by only including com here an not the module
header files (which in turn includes WebSockets, CryptAlgo)