Wayland: Add xdg-shell protocol files
Previously we linked with libweston-desktop. However the symbol
is no longer present in libweston-desktop8.so, it seems maybe
by chance we were lucky with previous versions.

I note that weston example clients include the protocol files
directly even when linking with libweston library, so we do
the same here.

Note: we should really depend on wayland-protocols and build
these protocol files from xml via cmake. But for now we include
the already generated files so as not to break buildroot
environment.

Signed-off-by: Peter Griffin <peter.griffin@linaro.org>