[ISSUES] Add functionality as requested in #216 and #217
As the functionality was added, it was observed that the config for the Plugin and the config for the Server
where intertwined, which is architecturally not a dependency, so refactored to seperate the two config.
While working on this, it was observed that the environment substitution does *not* take the callsign namespace
into account in case the substitution is part of a plugin config. Added this functionality.
As a preperation for isssue #154, where we would like to "configure" the path to use callsign or classname to
group persistent space into a sharing space, if required. Follow up for this issue is required after this
commit is merged.