[INTERFACES] Due to the windows build, the interface generation is more critical.
Since we now have EXTERNAL keyword applied to "all" interfaces,
it means that the default methods, Default constructor, default
destructor should be in an interface file so (definitions.so).
However, the windows build requires a dllimport and dllexport
on the right locations.

Due to these requried changes, it is observed that the current
dependencies for the definitions artifact
(contains all enums conversions exposed for the interfaces)
was depending on COM (and thus tracing) which is a needless dependency.
To overcome this dependency, the required COM interfaces
(and plugin interfaces, but these where already decoupled)
should not pull in any dependencies on COM.

This commit removes these dependencies.