[COM] Bartjens law reference count adjustment
This commit adjusts the missing behavior in the 'Bartjens law' case.

Previously, the following case was not handled:
	* Some trigger(A) queries the interface(X) across process boundaries
		and does not release it within the same function call.
		Instead, it stores it and releases it only on X's deactivation.
		During this, another trigger(B) queries the same interface
		and releases it in the same function call.

If both of those queries came from the same process and both went over
the process boundary, the second release did not decrement the
implementation reference count.

Currently, if the trigger B fires the QueryInterface, an internal
proxy counter is incremented. This counter signifies how many remote
releases have to be done, when the proxy gets cleaned up from a process
space.