[BUILD] pragma once does not prevent against copy of this file in other locations :-)
Reverting the usage of #pragma once in this file as for some reason the IShell gets copied for the ProxyStubs/JSON generated sources.
However due to this copy, the pragma once is assuming thta these are different files and I get, during the building of the ProxyStubs
a double definition error. Prevented this error by reverting to the preprocessor inclusion paradigm!