[CALLSTACK] Callstack dumping functionality fixed.
Since the software generation is now with a stricter compiler set:
1) No exception handling
2) Default hidden unless marked as public

The callstack dump on the CLI was broken. Turning of exception handling also
turns of the generation of unwind tables. Currently, even if exception handling
is off, the generation of the unwind-tables is enabled. Only for productions
builds this is also turned off.
During analysing the issue, found that the DumpCallStack and PublishCallstack
had overlapping functionality. Moved it to 1 procedure.