[Plugins] Added PluginInterfaceType
PluginInterfaceType is a generic helper class, which is meant to
wrap the functionality of subscribing to a StateChange notification
of a specific plugin.

In order to use it, one has to inherit from this class and implement
the Activated & Deactivated methods. Those are called upon receiving
the corresponding state change notification, a specified plugin.

Example usage:

```

class PlayerInterface
	: public PluginHost::PluginInterfaceType<Exchange::IPlayer>
{
	// override Activated & Deactivated
}

// ...

PlayerInterface _playerInterface;
_playerIterface.Activate(PluginHost::IShell*, "Callsign");

```