[WPEFRAMEWORK] Make sure HTTP requests can be made a-synchronous
The WebBridge plugin delegates incoming requests to the Javascript plugin running in a JavaScriptInterpreter. This can potentially take
some time so the system decouples the request from the response to the Javascript plugin.
In case an HTTP request with a JSONRPC message is received, it was expecting a response on the invoke. Due to the decoupled WebBridge
feature a JSONRPC request now nolonger return a response by default.

This patch assumes that if there is no JSONRPC message response on the invoke that it is an a-synchronous call and response will follow
later. As  part of the initial request, is the request to "close" the TCP connection on completeion of this request, this information
is forwarded and stored now as part of the channelInformation. If finally the JSONRPC response is received (and wrapped into an always
HTTP response 200 OK message), the channel is closed (if requested) after the JSONRPC response has been send.