[PROCESSNAME] Change the WPEProcess -> Callsign.
Change the WPEPross name to the callsign. Callsigns like org.rdk.webserver will be
abbreviated to the last work after the last dot only (so org.rdk.webserver -> webserver).

Reason for this is that the best supported conversion is only for the short process naem wich is limited to 16 characters.