[WPEFramework] Fix access permission for /tmp/keyhandler
Some of the process (like Amazon plugin) can run with a different
user and group than the Framework main process. In this
case, the processes that are using /tmp/keyhandler can't access
that file, thus loosing possibility to obtain key presess.
This commit fix this, with adding R/W file permission for
the 'other' group.