Bluetooth: GATTSocket: Don't set SocketPort send/recv sizes based on maxMTU
The GATTSocket constructor is using its 'maxMTU' parameter for setting
the CommandSink's preferred MTU, but also for setting the send/recv
buffer sizes from on the underlying socket managed by SocketPort. These
are different concepts and should not be mixed, because the socket
buffer sizes should be large enough to accommodate for send/recv bursts
of several MTUs and latency on socket polling.

This actually causing incoming packet loss on the GATTSocket, e.g.,
when used from the BluetoohRemoteControl plugin, which is setting the
'maxMTU' to 255 bytes (kernel then imposes a 2240 bytes minimum),
resulting in RCU key events and voice data packets loss.

This way, change the GATTSocket constructor implementation to let
SocketPort (and then the kernel) use the defaults for the socket buffer
sizes, by specifying ~0 for both send and receive buffer sizes.

To still allow configuring the actual socket buffer sizes, introduce a
new GATTSocket constructor with arguments for the socket's send and
receive buffer sizes.

Signed-off-by: Ricardo Silva <ricardo.josilva@parceiros.nos.pt>