Bluetooth: Only set L2CAP_LM option if setting BT_SECURITY option fails
When setting the GATTSocket Security, the sequence should be first to
try and set the SOL_BLUETOOTH/BT_SECURITY option first and, only if that
fails, try setting the SOL_L2CAP/L2CAP_LM equivalent option. This is
what the reference code (bluez's gattool) is doing.

This way, rework the option setting logic in GATTSocket::Security() to
follow what the reference code does.

Signed-off-by: Ricardo Silva <ricardo.josilva@parceiros.nos.pt>