Make sure to export cmake modules outside Thunder
Previously, cmake/modules were not installed and not added to the
CMAKE_MODULES_PATH. This resulted that external components,
that are using find_package(WPEFrameowrk), couldn't acccess common modules.

This is now changed, so that Thunder will install the modules in the
sysroots /usr/lib/cmake/WPEFramework. It will export that path,
so the find_package(WPEFramework) will append CMAKE_MODULE_PATH
with module location alowing usage of the modules.