[ONEM-22696] Fix for memleak on each socket connection.
In case the IdleTimer is set to 0 (no checking on dead connections) the created connections, never get cleared.
This patch will clear closed sockets on every new connection.
During testing we found that if the ProxyType during casting requires an invalid proxy, it ASSERTS. Now checking
for an invalid proxy before we cast.