[REFERENCELEAK] The new ProxyTypes did not release an existing reference if the asignment operator overload was triggered. (#798)
This causes a small memory leak that could only be detected in case the memory graph got plottted. To prove the stability of the
proxypool, the WPEFramework application can now print the elements created in the factories as used by the WPEFramework plugin
server (press E) under continous load these should be stable.