Virtual Input: Import/Export keymap from/to JSON array
Keymaps are curently loaded/saved using methods that imply that a dedicated file
exists to store the keymap. That is not always the real use case, since this map
can be embedded on a file containing other types of data.
For the above reason, this commit adds new methods to Import/Export the keymap
by parsing/filling a JSON array of 'KeyMapEntry' data types.
In order to re-use code:
 * 'VirtualInput::Keymap::Load' calls the new 'VirtualInput::KeyMap::Import'
 * 'VirtualInput::Keymap::Save' calls the new 'VirtualInput::KeyMap::Export'
By having these new methods, we can consider that 'VirtualInput::Keymap::Load'
and 'VirtualInput::Keymap::Save' can be considered DEPRECATED.