Build: Export CRITICAL_SECTION_LOCK_LOG compile setting
While having the DEADLOCK_DETECTION CMake option turned on, an 'undefined symbol'
error occurred when loading a shared library on a plugin.
Comparing the generated libraries, there was an obvious difference in the method
signature. This is caused by the missing declaration of CRITICAL_SECTION_LOCK_LOG
compilation specifier, on the plugin side.

CRITICAL_SECTION_LOCK_LOG compilation specifier changes the implementation on
CriticalSection, thus creating type misalignement when is not declared on both
sides. Currently it is not being exported.