[FRAME] Extend the functionality of the Frame template. (#840)
* [FRAME] Extend the functionality of the Frame template.
Sometimes the ordering in a binary serialzed frame is LSB, sometimes it is MSB. Allow the template to
cope with both situations, the default is MSB as that was as it was used sofar.
Also sometimes we only need to work with very small frames <255 bytes. using the the uint16_t for maintaining
location and size is a bit of overkill. Maybe in the future we would like to enlarge the 64Kb boundary to
4Gb, and the variables than need to be uint32_t :-)
Make the variable sizes for creating the buffers a template parameter, again making the default equal to
what it was.

* [FRAME] Update after review.

[WINDOWS] Make new Time compile with less warnings on Windows.

Co-authored-by: MFransen69 <39826971+MFransen69@users.noreply.github.com>