[WORKERPOOL] Refactor of the workerpool to close all loopholes.
The current implementation of the Workerpool does allow for a Job to be running on
multiple threads of the workerpool. Thuis is due to the fact that in the current
implementation the Job can be resubmitted once it is started.
This means that if a job is running, during this time the Job can be scheduled
for execution again. Which means that the job could already start/run on another
thread while the first instance is still running.

This is conceptually not correct and might potentially cause starvation of the
threadpool (what if the job never ends due to a bug or takes very ong to execute
and gets submitted in a short tiem frame periodically).

To prevent this starvation, the job now has a more eleborated states. See:
core/ThreadPool.h class JobType::state.