DecoupledJob: Move DecoupledJob class to WorkerPool.h
DecoupledJob class is better suited to be placed as a common component in the
framework core. It will also facilitate maintenance, since the implementation is
copied throughout several packages.
In the process of moving it from Bluetooth plugins to WorkerPool.h file, a fix
is introduced that will facilitate the destruction process of an object of this
type. Original implementation has a caveat when a Job is added to the ThreadPool
MessageQueue and the submitter process is destroyed/deactivated in-between. For
this use case, if Revoke is not exercised, it can lead to misbehaviors.
With this new implementation :
 * the destructor was modified in order to call Revoke method instead of being
   left as default.
 * Submit method return code was changed from void to bool so that the caller
   can know whether the requested job was successfully submitted or not.