[DEADLOCK] Make sure the Destroy (shutting down of the WPEFramework) does not lock the ServiceMap.
The lock in the destroy was around the map, and in the scope of this lock the deactivation of the
service was called. Before this was oke but now, the Deactivation might require a cleanup of the JSONRPC
not unregistered connection. This requires (during deactivation, from the Monitor Thread) access to the
map again to signal a closed connection. Cuasing a deadlock at shutdown.

Moved the deactivation of the service outside of the map lock. That is possible as the services are refcounted
and thus a service list to deactivate is created within the lock (and ref counted) and than, ouside of the lock
deactivate all selected services.